# Using StringBuilder to improve perf when handling long strings