@sebastienros Please take a look.

## Goal
JSON.parse and JSON.stringify slow down a lot when handling long strings, #357 
Using StringBuilder.Append improves the performance a lot.

## Benchmark
|String.Length|Before(ms) |  After(ms) |
|---|---|---|
|10|0 |0 |
|100|0 |0 |
|1,000|1 |0 |
|10,000|22 |1 |
|100,000|2166 |7 |
|1,000,000|(too long) |74 |
### Test Code
```

namespace ConsoleApp1
{
    using Jint;
    using System;
    using System.Diagnostics;
    using System.Linq;

    public class Program
    {
        static void Main(string[] args)
        {
            var repeatCount = 1;
            Parse(repeatCount);
            if (args.Length > 0)
            {
                repeatCount = int.Parse(args[0]);
            }
            Stopwatch sw = new Stopwatch();
            sw.Start();
            Parse(repeatCount);
            sw.Stop();
            Console.WriteLine(repeatCount + " times stringify takes: " + sw.ElapsedMilliseconds + " milliseconds.");
        }

        private static void Stringify(int repeatCount)
        {
            Engine engine = new Engine();
            var data = GetTestData(repeatCount);
            engine.SetValue("a", data);
            engine.Execute("JSON.stringify(a)");
            var k = engine.GetCompletionValue();
        }

        private static void Parse(int repeatCount)
        {
            Engine engine = new Engine();
            var data = GetTestData(repeatCount);
            var stringified = "{ \"s\": \"" + data + "\"}";
            engine.SetValue("a", stringified);
            engine.Execute("JSON.parse(a)");
            var k = engine.GetCompletionValue();
        }

        private static string GetTestData(int repeatCount)
        {
            const string data = @"0123456789";
            return string.Concat(Enumerable.Repeat(data, repeatCount));
        }
    }
}

```